move library and search model into ESMapping remove libraries get and just use the post with search params make from and size actually optional use more granular case classes to interpret the json body
